Further implementation of Avahi DNS-SD functionality
There's a problem with structure and ownership of the data
by objects.
In case of Avahi implementation, Discovery object is passed an
exclusive ownership (unique_ptr) of the client_p, which functions
as a context pointer to the session. When it dies, the client
disconnects - is not browsable, and we cannot perform browsing
or any action related to a client (service).
This should not be managed by a discovery object,
as discovery object performs only DNS-SD functionality
with written logic that returns a collection of browsed service
data.
Moreover, the discovery thread once started, cannot be resumed
as no class has access to discovery object.
This will be a problem once the client disconnects, and we want
to reconnect.
Either BVApp class egts extended responsibilities,
or new object should be created (a manager or something).
Changes:
+ Information about bridged connection between VM and host
+ Moved discovery constants to BVDiscovery
+ BVServiceHostData struct for holding the data
  initialized by the BVService class.
  It should be passed to BVDiscovery instead of pointer to the
  BVService. This is yet another issue steming from the stucture.
+ Discovery object for Avahi WIP
+ Misc notes